add cross-platform test and release 0.4.1 (#331)
* bump to 0.4.1

* add cross-platform test

+++ test/base/env.py
@@ -73,10 +73,10 @@ class MyTestEnv(gym.Env):
         elif self.recurse_state:
             return {'index': np.array([self.index], dtype=np.float32),
                     'dict': {"tuple": (np.array([1],
-                                       dtype=np.int64), self.rng.rand(2)),
+                                       dtype=int), self.rng.rand(2)),
                              "rand": self.rng.rand(1, 2)}}
         elif self.array_state:
-            img = np.zeros([4, 84, 84], np.int)
+            img = np.zeros([4, 84, 84], int)
             img[3, np.arange(84), np.arange(84)] = self.index
             img[2, np.arange(84)] = self.index
             img[1, :, np.arange(84)] = self.index
